requirements.txt
nltk == 3.7
python-dotenv